requirements.txt
aiogram==2.23.1
httpx==0.28.1
Pillow==11.3.0
